feat(frontend): refactor selector for better maintenance and readability
