+++ packages/frontend/src/components/selector/hooks.ts
@@ -1,0 +1,19 @@
+import { useEffect, RefObject } from 'react'
+
+export const useOutsideClick = (
+  ref: RefObject<HTMLElement | null>,
+  callback: () => void
+) => {
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (ref.current && !ref.current.contains(event.target as Node)) {
+        callback()
+      }
+    }
+
+    document.addEventListener('mousedown', handleClickOutside)
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+  }, [ref, callback])
+}
